+++ app/api/routes.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, HTTPException
 from app.dto.schemas import CorrectionRequest, CorrectionResponse, ErrorResponse
 from app.services.correct_text_helper import process_correction_request
 from app.services.feedback_generator import generate_feedback
-from app.utils.logger import get_logger
+from app.common.utils.logger import get_logger
 
 logger = get_logger(__name__)
