+++ app/clients/alibaba/llm/qwen_client.py
@@ -4,7 +4,7 @@ from openai import OpenAI
 from app.config.app_config import settings
 from app.consts.prompts import get_correction_prompt
 from app.dto.schemas import CorrectionResponse
-from app.utils.logger import get_logger
+from app.common.utils.logger import get_logger
 
 logger = get_logger(__name__)
