+++ app/services/correct_text_helper.py
@@ -1,7 +1,7 @@
 from app.clients.alibaba.llm.qwen_client import correct_text
 from app.services.pattern_matching import validate_correction
 from app.services.feedback_generator import generate_feedback
-from app.utils.logger import get_logger
+from app.common.utils.logger import get_logger
 
 logger = get_logger(__name__)
