+++ app/services/feedback_generator.py
@@ -1,5 +1,5 @@
 from app.dto.schemas import CorrectionResponse
-from app.utils.logger import get_logger
+from app.common.utils.logger import get_logger
 
 logger = get_logger(__name__)
